Pull out reusable test components
